[REF] contract: Remove duplicated O2m
Equirenta only needs the contract lines view that is used when the
`line_recurrence` field is `False` so we are removing the field
`contract_line_fixed_ids` to simplify the models and because having both
fields made some computation fields being miscalculated, apparently
they were sharing the cache between both fields and when showing them
the calculation goes to the field that is hidden.

Also, adding the attribute `change_default` to the `product_id` because
that field triggers an `onchange`.